+++ demo-app/src/lib.rs
@@ -1,0 +1,13 @@
+#[cfg(test)]
+mod data_generation;
+#[cfg(test)]
+mod helpers;
+mod runtime;
+#[cfg(test)]
+mod test_utils;
+#[cfg(test)]
+mod tests;
+mod tx_hooks_impl;
+#[cfg(test)]
+mod tx_revert_tests;
+mod tx_verifier_impl;
